conversationmodel: fix contact selection for interactions and removal
Consider a conversation's read-only attribute when querying for indices
that match a given uri. This is done when searching for a conversation
associated to an interaction(in the rare case that more than one 1to1
conv exists).

Also remove all the conversations for a contact once the contactRemoved
slot is called.

Gitlab: #475
Change-Id: I637dbcf677a5e00b6b52320c3cd24fcd08cfacd6