video: do not start unused renderers
Avoid a lot of unnecessary work in the daemon.

Note: This is a hack and should be reverted in the future. However, this
needs major changes in the daemon's API to rewrite the current video stack
here, and video-split is needed. So, for now use a whitelist in callmodel
to know when registering. In the future the client will register and unregister
the renderer when needed

GitLab: https://git.jami.net/savoirfairelinux/jami-daemon/-/issues/658
Change-Id: Ibf581cca3a4bf557502687a68b7ab5989d11c38a